日報用コントローラの実装先をミスっていたため新規コントローラ作成 + 内容移行
