+++ resources/views/user/daily_report/show.blade.php
@@ -25,7 +25,7 @@
   <div class="btn-bottom-wrapper">
     <a class="btn btn-edit" href="{{ route('daily_report.edit', $report->id) }}"><i class="fa fa-pencil" aria-hidden="true"></i></a>
     <div class="btn-delete">
-      {{ Form::open(['route' => ['daily_report.delete', $report->id], 'method' => 'DELETE']) }}
+      {{ Form::open(['route' => ['daily_report.destroy', $report->id], 'method' => 'DELETE']) }}
         <button class="btn btn-danger" type="submit"><i class="fa fa-trash-o"></i></button>
       {{ Form::close() }}
     </div>
