+++ routes/web.php
@@ -30,12 +30,14 @@ Route::group(['prefix' => '/', 'user.', 'namespace' => 'User'], function () {
 
 
     Route::get('home', 'UserController@index')->name('home');
-    Route::get('user/daily_report/create', 'UserController@create')->name('daily_report.cerate');
-    Route::post('store', 'UserController@store')->name('daily_report.store');
-    Route::get('user/daily_report/{id}/show', 'UserController@show')->name('daily_report.show');
-    Route::get('user/daily_report/{id}/edit', 'UserController@edit')->name('daily_report.edit');
-    Route::put('user/daily_report/{id}', 'UserController@update')->name('daily_report.update');
-    Route::delete('user/daily_report/{id}', 'UserController@delete')->name('daily_report.delete');
+
+    Route::get('report', 'ReportController@index')->name('report');
+    Route::get('user/daily_report/create', 'ReportController@create')->name('daily_report.cerate');
+    Route::post('store', 'ReportController@store')->name('daily_report.store');
+    Route::get('user/daily_report/{id}/show', 'ReportController@show')->name('daily_report.show');
+    Route::get('user/daily_report/{id}/edit', 'ReportController@edit')->name('daily_report.edit');
+    Route::put('user/daily_report/{id}', 'ReportController@update')->name('daily_report.update');
+    Route::delete('user/daily_report/{id}', 'ReportController@destroy')->name('daily_report.destroy');
 
     Route::get('attendance', ['as' => 'attendance.index', 'uses' => 'AttendanceController@index']);
     Route::post('attendance/register', ['as' => 'attendance.register.start', 'uses' => 'AttendanceController@setStartTime']);
